Move default slash commands to respective module (#4240)
* Move default slash commands to respective module

* Update public/scripts/slash-commands.js


---------
